moved cargo and hatch commands to new subpackages
